Add files via upload
Command line utility to lock and unlock the ability for Forza Motorsport 7 to communicate UDP packets on 127.0.0.1 (localhost). This is provided as a community service and is not needed if you own a SimXperience product and therefore use Sim Commander version 4 or higher.


Microsoft .Net Framework version 4.5 or higher is required to be installed for this utility to function properly. NOTE: This tool has been tested to work correctly only on my development machine. In theory it will work as intended on other machines. Nonetheless, please consider this software to be experimental and USE AT YOUR OWN RISK!!!